filter: Move write_vcf() and _filename_gz() to augur.utils
These functions don't have filtering logic and can be moved next to the existing read_vcf() as a more appropriate scope.
